# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduSaaS - Plataforma Educacional Multi-tenant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in current tenant
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, coordinator, teacher, student]
        - name: classId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      summary: Get user by id
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      summary: Delete user
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{userId}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get student performance stats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentStats"

  # ─── ACADEMIC STRUCTURE ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH ──────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, coordinator, teacher, student]
        avatarUrl:
          type: string
          nullable: true
        registrationNumber:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tenantId, name, email, role, createdAt]
...
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, coordinator, teacher, student]
        registrationNumber:
          type: string
          nullable: true
        classId:
          type: integer
          nullable: true
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, coordinator, teacher, student]
        registrationNumber:
          type: string
          nullable: true
    StudentStats:
      type: object
      properties:
        userId:
          type: integer
        totalExamsTaken:
          type: integer
        averageScore:
          type: number
          nullable: true
        bestScore:
          type: number
          nullable: true
        bySubject:
          type: array
          items:
            $ref: "#/components/schemas/SubjectPerformance"
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
      required: [userId, totalExamsTaken, bySubject, recentSessions]

    # ─── ACADEMIC ──────────────────────────────────────────────────────────────
...
    SessionSummary:
      type: object
      properties:
        sessionId:
          type: integer
        examId:
          type: integer
        examTitle:
          type: string
        examType:
          type: string
        score:
          type: number
          nullable: true
        percentage:
          type: number
          nullable: true
        submittedAt:
          type: string
          nullable: true
        status:
          type: string
      required: [sessionId, examId, examTitle, examType, status]
...
    SubjectPerformance:
      type: object
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        averageScore:
          type: number
          nullable: true
        totalAttempts:
          type: integer
        color:
          type: string
          nullable: true
      required: [subjectId, subjectName, totalAttempts]
...